Changes record layout to hold 2^48 pointer sizes
as well as room for version fields (for MVCC) and at the same time
aligning with 2^N sizes, e.g. 32, 64, 128 a.s.o.

This is to be considered a spike to enable a store to grow to a very large
number of entities. No consideration w/ regards to minimizing store size
overhead have been taken at all.