Allows for more fine grained InputEntity decoration
Basically flips the decoration of input entities inside out, so that the
decoration happens while building the entity, before materializing it.
That way a decorator can have better control over how the entity will look
like, and there is less need to expose ways of altering the created entity
afterwards.

A con is that also the "public" decorators, namely additive labels and
default relationship type, needs to be injected into the Input instance and
therefore forces an Input into using the Builder (previously known as
Deserialization) for building input entities.