Lenient record reading when traversing relationship chains
StoreRelationshipIterable is the replacement for RelationshipChainLoader
from 2.2. RelationshipChainLoader used a more lenient way of loaing the
chain records, where finding an unused record didn't throw exception,
rather stop the loading of that chain.

This commit changes 2.3 to do also that, but takes things one step further
by trying to use the information within the unused (assumed recently
deleted) records to follow its pointer onto the next record. So
behaviour-wise this brings 2.3 closer to 2.2 in this aspect.