More informed selection of pivot in qsort in ParallelSort
Problem being that the qsort algorithm, which is implemented with
recursion, may be extremely unlucky and hit StackOverflowError before
reaching the leaves to start sorting.

Pivot base is (and was) selected at random. In addition to that this
commit adds so that 10 surrounding items are looked at and the mean
of those selected as the pivot to use. This should result in a better
worst-case and reduced probability of SOE.

When reading this commit it may look like much have chanced. That's
because a couple of methods have been moved into SortWorker where they
belong. The only real change is the additional call to choosePivot in
`recursiveQsort` method. That and removal of an unnecessary start signal
of all the sort workers.

On top of this the sorting algorithm will be retried a couple of times if
a StackOverflowError is still hit.